+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [[v0.4.7]](https://github.com/mlange-42/ark/compare/v0.4.6...v0.4.7)
+## [[unpublished]](https://github.com/mlange-42/ark/compare/v0.4.6...main)
 
 ### Performance